Extract Prompt Blink Animation hook from promptForm file
