+++ src/app/[locale]/dashboard/[id]/@promptForm/PromptForm.tsx
@@ -32,6 +32,7 @@ import AccordionDetails from '@mui/material/AccordionDetails';
 import { useRouter } from '@/navigation';
 import { useParams } from 'next/navigation';
 import Link from 'next/link';
+import usePromptBlinkAnimation from './usePromptBlinkAnimation';
 
 const totalRows = 12;
 const totalColumns = 12;
@@ -67,43 +68,6 @@ function SubmitButton({ text }: { text: string }) {
   );
 }
 
-const usePromptBlinkAnimation = (
-  prompt: PromptRecord,
-  setControlledPromptValue: React.Dispatch<React.SetStateAction<PromptRecord>>,
-  preventBlink: React.MutableRefObject<boolean>,
-) => {
-  const [blink, setBlink] = React.useState(true);
-  const trigAnimation = React.useCallback(() => {
-    if (prompt) setBlink(false);
-    setTimeout(() => {
-      setBlink(true);
-      setControlledPromptValue(prompt);
-    }, 300);
-  }, [prompt, setControlledPromptValue]);
-
-  const useTrigAnimationOnPromptChange = ({
-    prompt,
-    trigAnimation,
-  }: {
-    prompt: PromptRecord;
-    trigAnimation: () => void;
-  }) => {
-    const prevPrompt = usePrevious(prompt);
-    React.useEffect(() => {
-      const promptIsUpdated =
-        JSON.stringify(prevPrompt) !== JSON.stringify(prompt);
-      if (promptIsUpdated && !preventBlink.current) {
-        trigAnimation();
-      }
-      if (preventBlink.current) preventBlink.current = false;
-    }, [prompt, trigAnimation, prevPrompt]);
-  };
-
-  useTrigAnimationOnPromptChange({ prompt, trigAnimation });
-
-  return blink;
-};
-
 const useFormStateWatcher = () => {
   const [state, formAction] = useFormState(submit, null);
   const [setBoard, setErrorOnBoardGeneration, stashDashboard] = useBoundStore(
@@ -123,14 +87,6 @@ const useFormStateWatcher = () => {
     }
   }, [state, setErrorOnBoardGeneration, stashDashboard, router, setBoard]);
   return formAction;
-};
-
-const usePrevious = <T,>(value: T): T | undefined => {
-  const ref = React.useRef<T>();
-  React.useEffect(() => {
-    ref.current = value;
-  });
-  return ref.current;
 };
 
 export function PromptForm() {
